requirements.txt

django==4.2.7

bs4==0.0.1
pillow==10.1.0
django-restframework==0.0.1
postman==0.6.0
django-widget-tweaks==1.5.0
django-crispy-forms==2.1
crispy-bootrap5==2023.10

google-api-python-client==2.108.0